Rename uint:bit modules
