+++ src/uint/bit_xor.rs
